convert function_column
